RB Developer Tools seed.
The main goal of the plugin is to provide a normalized way of generating
custom fields that allows to modify meta data values, but with the intention
of being adaptable to any other kind of value (forms for example) if needed.

The fields generator is based on `React`. Its **detachable and can be used
in any project if wanted**. It follows a **single, group, repeater model**

- Single: component that manages single value changes. It can be any of the
componentes from the `wp.components` set, or a custom one.
- Group: set of single, groups, repeaters. Stores the values in a JSON
- Repeater: array of single, group or repeaters. The user manages the
amount of items stored.

On top of the fields generator there will be **components that positionate
the fields according to the object type** and config passed (post type, term, etc).
At the end, the should be PHP functions that allow to add this fields to
post, terms, menu items, settings page, attachments page and modals, and customizer panel.

At this point, there is an stable set of functionalities that allow to stablish
post meta value fields that appear on the gutenberg editor and modifies the
current post meta correctly.

**This functionalities are abstracted to PHP**, meaning that a wordpress dev
wouldn't need to learn `JS` or `React` to add this metafields.

Another goal is to integrate it with a translation plugin. Maybe by appending
language suffixes to the meta names, and adding an easy way (context or redux)
to modify the meta key being modified by the control, so it changes from, for example,
`meta_key_es` to `meta_key_en` accordingly.