- Common functionalities used between meta fields managers (post and terms for now) moved into the `RB_Meta_Field` trait. A trait was used instead of an abstract class because it allows us to use static methods and properties that are linked to the class that uses the trait. With an abstract approach, the static properties refer to the abstract class and not the class that is implementing it.
- Change property name `fields` to `field` in the `add_field` method.
Even though multiple fields can be added to control a meta value in the
form of repetear and groups, in the end, it is one component that manages
the final value of the meta. So `add_field` now receives `field`, but the
child filds still receive `fields`, as it determinates the field type.

- Added `RB_Term_Meta_Field` seed. As it is, it only registers fields,
but there is no front code using them.

- Added trait `Initializer` to manage static classes initializations.

- Changes styles of repeaters. Removed background color and added black
borders.