Major changes: ------------------------
- Modified the pattern used to manage and create meta fields controllers.
Now, the new `RB_Meta_Fields_Manager` creates an instance that manages
the fields of a certain `object_type` passed through `__construct` with other
arguments.
`RB_Post_Meta_Field` and `RB_Term_Meta_Field` implements an instance of
`RB_Term_Meta_Field` instead of inheriting.
The `RB_Meta_Field` is used in the classes to provide the common methods.

- `RB_Term_Meta_Field` has an almost complete implamantation of the fields
system in the term edit form.
It generates a placeholder for the meta row that gets filled by the
`rb-term-edit-form-fields`, and it manages the save proccess of the new
fields using the `WP_REST_Terms_Controller` and `WP_REST_Term_Meta_Fields`,
that handle the update of meta values based on the registered schema.

- Added `rb-term-edit-form-fields` script, that attaches the registered
term meta fields to the term edit form.

Minor changes:
------------------------

- Added `RBFinalField` component that renders an `RBField` and handles
the value change.

- `RBField` first parent `depth` changed from `1` to `0`.

- `RBField` hidden input containing the stringified value is not rendered
in depths deeper than 0. This is to prevent conflict between the inner
fields of the meta value field and any other fields in the form.

- `RepeaterField` background change from `transparent` to `white` to simulate
the style from the editor in the term edit form.

- Created the `parsePHPFieldData` and `parseObjectPropNames` helpers based
on code that use to be present in `rb-post-meta-fields` that is now used to
by `rb-term-edit-form-fields`.

- Added test files that add meta fields for post and terms.