Added support for RB Fields in the media popup.
- Attachment fields can now be modified in the media popup.
Instead of filtering the $form_fields via `attachment_fields_to_edit`,
the backbone api of the media popup has been extended to generate a space
that renders the custom fields. This way we can manage correctly the data
fetch and save triggers, and most importantly, the render and unmount of
the react component from the DOM.

- In `RB_Post_Meta_Fields_Manager` the attachment data sent through the
REST API gets filtered to include the `rbfields` property, that includes the
`fields`, `placeholder`, and `values` fields, that are used to generate
the fields in media popup using the `rb-media-popup-fields` script.

- Added the `rb-media-popup-fields` that gets enqueued when the media
scripts are enqueued using `wp_enqueue_media`.

This script extends the backbone api of the media popup to manage
render, unmount and save events of the custom fields.

- In `RBField`, the main field input (RBField of depth 0) gets
it change event triggered when the value passed by props changes.

This makes the field compatible with other js apis that work listening
to the input change, like the customizer panel, or the media popup attachment
metas.

- Added method `RB_Meta_Fields_Manager::get_subtype_fields` that return
the meta fields registered to a certain subtype.