Wordpress object meta field column creation automated for terms and posts.
- Fields REST route `rb/$namespace/v1/$object_subtype` changed to `rb-fields/v1/$object_subtype/`.
This allows to make the route more dynamic, being able to use it only knowing the
`$object_subtype`.

- Added initialization to `RB_Meta_Fields_Manager` that enqueues the `rb-object-list-column-fields`
script that renders the fields in the wordpress objects list field placeholder
column.

- Added the trait `RB_WP_Object_Field` used by `RB_Post_Meta_Field` and `RB_Term_Meta_Field`
to stablish the variables and processes that are common to any wordpress object type.
This trait sets up the objects list field column, so the logic that was in `RB_Post_Meta_Field`
in the last commit was moved into it, to make it easy to implement in the `RB_Term_Meta_Field`.

- Changed component `PostMetaField` name to `GutenbergPostMetaField` to better
reflect the dependency this component has to the gutenberg editor.

- Added the `WPObjectMetaField` component that renders a field that modifies the
value of a wordpress object meta.

- Script `rb-object-list-column-fields` modified to correctly render the
`WPObjectMetaField` over the placeholder based on the variables on `RBObjectsList`.

- Fixed bug in the columns setup that caused custom field column with a position
different that the last one possible to break.