Further refactorization of the RB Customizer classes.
- `RB_Customizer_Control` now only accepts one setting. Wordpress doesn't
allow to link a control to more than one setting, some having it be an array
of `setting_id => setting_config` was confusing, as it only allows one.

The setting id can be passed through the setting config, but if none is
provided then the control id will be used when registering the setting.

- The `RB_Customizer_Section` now adds its selective refresh partials on the
`customzer_register` action with a absurdly high priority to make it run after
all the controls and settings have been added.

Before it was doing it on `__construct`, which would cause it to not register
anything, as the controls and settings haven't been defined yet.