RB Fields user edit meta implementation.
- Added `RB_User_Meta_Fields_Manager` to manage the creation of fields
for users. It is a little different from other `RB_Object_Type_Fields_Manager`
implementations because the user object type doesn't have any subtypes, so
fake subtype slug and kinds had to be hardcoded to make it work with the
`RB_Object_Type_Fields_Manager` functionalities.

- Added `RB_User_Meta_Field` to render the placeholders of the registered
user fields in the correct places.

- Added the function `rb_update_user_meta` that utilizes wordpress user
REST controllers to update the meta based on the registered schemas.

- Added the script `rb-user-fields` that renders the REACT fields on top
of the user fields placeholders.

- `RB_Object_Type_Fields_Manager::get_field_subtypes` now returns an array
containing as only item the default kind if a kind wasn't passed through
the `$field_args` parameter.