Implemented fields in the quick edit box for objects lists.
The only weird thing with this implementation is that since there is no way
to listen to the quick edit save event, we listen to it using the `ajaxComplete`
jQuery event and checking for the action. This doesnt take in mind the fact
that even if the ajax request finished, it may have failed, thus not closing
the editbox.

Normally, after the quick edit box is saved, Wordpress fetches the row with
the updated values for the columns and renders it. Since the column fields
uses the WP JS API to access and modify the `getEditedEntityRecord` values,
when the re-render happen, if the values haven't been updated with the new values
from the quick edit box, they will be using the old values previous to the quick
changes.

Should we not used `getEditedEntityRecord`? or check if we can update
`getEditedEntityRecord` after the field re-render instead of save?

Changes and features
----

- `RB_WP_Object_Field` now can recieve the `quick_edit` setting that stablishes
if this field should be rendered in the object quick edit box in the objects lists table,
and rendering it accordinly. The classes that uses `RB_WP_Object_Field`

- Added the abstract method `RB_WP_Object_Field::is_own_quick_edit` to determine
if we are on a quick edit box for this object type.

- In `RB_WP_Object_Field` the field data attributes are now rendered on
a parent element on top of the placeholder to make this data available even
if the field render is a custom callback instead of a field placeholder for the
JS API.
This way the quick edit field can access the initial meta value and update
it accordingly even if the field render in the column is a custom callback.

- Extended the `rb-object-list-column-fields` to render the quick edit
fields and update the column fields values on quick edit changes.