Added the option to hide columns when adding a quick edit field.
Quick edit fields only can be rendered for every custom column added. In order
to hide this column if we don't want it, the option `hide_column` has been added
to the `quick_edit` option.

- Added the `RB_Object_Type_Fields_Manager::generate_field_instance($field_data)`
method to generate an instance of `RB_WP_Object_Field` when adding a new field, and
store it in the static `fields` property.

This way we can access every instance of `RB_WP_Object_Field` generated and filter,
for example, the hidden columns for a kind based on the `is_column_hidden`
and `get_field_config_kinds` methods of `RB_WP_Object_Field`.

- Added the `RB_Object_Type_Fields_Manager::get_kind_hidden_columns($kind_name)`
method. Returns the hidden custom columns for a kind if any.

- The `RB_WP_Object_Field` implementations and `RB_Menu_Item_Field` no longer receive
the field arguments because this are already passed and sanitized into `field_config`.